Reverting address back to int64 because of sql null types

+++ internal/endpoint/http/requests/user/user_request.go
@@ -1,7 +1,7 @@
 package request
 
 type UserRequest struct {
-	Address   int    `json:"address"`
+	Address   int64  `json:"address"`
 	FirstName string `json:"firstName"`
 	LastName  string `json:"lastName"`
 }
